+++ src/JsonSchema/Constraints/BaseConstraint.php
@@ -39,18 +39,15 @@ class BaseConstraint
      */
     protected $factory;
 
-    /**
-     * @param Factory $factory
-     */
     public function __construct(?Factory $factory = null)
     {
         $this->factory = $factory ?: new Factory();
     }
 
-    public function addError(ConstraintError $constraint, ?JsonPointer $path = null, array $more = [])
+    public function addError(ConstraintError $constraint, ?JsonPointer $path = null, array $more = []): void
     {
-        $message = $constraint ? $constraint->getMessage() : '';
-        $name = $constraint ? $constraint->getValue() : '';
+        $message = $constraint->getMessage();
+        $name = $constraint->getValue();
         $error = [
             'property' => $this->convertJsonPointerIntoPropertyPath($path ?: new JsonPointer('')),
             'pointer' => ltrim(strval($path ?: new JsonPointer('')), '#'),
@@ -76,7 +73,7 @@ class BaseConstraint
         $this->errorMask |= $error['context'];
     }
 
-    public function addErrors(array $errors)
+    public function addErrors(array $errors): void
     {
         if ($errors) {
             $this->errors = array_merge($this->errors, $errors);
@@ -89,7 +86,10 @@ class BaseConstraint
         }
     }
 
-    public function getErrors($errorContext = Validator::ERROR_ALL)
+    /**
+     * @phpstan-param int-mask-of<Validator::ERROR_*> $errorContext
+     */
+    public function getErrors(int $errorContext = Validator::ERROR_ALL): array
     {
         if ($errorContext === Validator::ERROR_ALL) {
             return $this->errors;
@@ -102,7 +102,10 @@ class BaseConstraint
         });
     }
 
-    public function numErrors($errorContext = Validator::ERROR_ALL)
+    /**
+     * @phpstan-param int-mask-of<Validator::ERROR_*> $errorContext
+     */
+    public function numErrors(int $errorContext = Validator::ERROR_ALL): int
     {
         if ($errorContext === Validator::ERROR_ALL) {
             return count($this->errors);
@@ -111,7 +114,7 @@ class BaseConstraint
         return count($this->getErrors($errorContext));
     }
 
-    public function isValid()
+    public function isValid(): bool
     {
         return !$this->getErrors();
     }
@@ -120,7 +123,7 @@ class BaseConstraint
      * Clears any reported errors.  Should be used between
      * multiple validation checks.
      */
-    public function reset()
+    public function reset(): void
     {
         $this->errors = [];
         $this->errorMask = Validator::ERROR_NONE;
@@ -129,22 +132,17 @@ class BaseConstraint
     /**
      * Get the error mask
      *
-     * @return int
      * @phpstan-return int-mask-of<Validator::ERROR_*>
      */
-    public function getErrorMask()
+    public function getErrorMask(): int
     {
         return $this->errorMask;
     }
 
     /**
      * Recursively cast an associative array to an object
-     *
-     * @param array $array
-     *
-     * @return object
      */
-    public static function arrayToObjectRecursive($array)
+    public static function arrayToObjectRecursive(array $array): object
     {
         $json = json_encode($array);
         if (json_last_error() !== \JSON_ERROR_NONE) {
@@ -160,22 +158,16 @@ class BaseConstraint
 
     /**
      * Transform a JSON pattern into a PCRE regex
-     *
-     * @param string $pattern
-     *
-     * @return string
      */
-    public static function jsonPatternToPhpRegex($pattern)
+    public static function jsonPatternToPhpRegex(string $pattern): string
     {
         return '~' . str_replace('~', '\\~', $pattern) . '~u';
     }
 
     /**
-     * @param JsonPointer $pointer
-     *
      * @return string property path
      */
-    protected function convertJsonPointerIntoPropertyPath(JsonPointer $pointer)
+    protected function convertJsonPointerIntoPropertyPath(JsonPointer $pointer): string
     {
         $result = array_map(
             function ($path) {