+++ src/JsonSchema/Constraints/ConstraintInterface.php
@@ -23,45 +23,37 @@ interface ConstraintInterface
 {
     /**
      * returns all collected errors
-     *
-     * @return array
      */
-    public function getErrors();
+    public function getErrors(): array;
 
     /**
      * adds errors to this validator
-     *
-     * @param array $errors
      */
-    public function addErrors(array $errors);
+    public function addErrors(array $errors): void;
 
     /**
      * adds an error
      *
-     * @param ConstraintError  $constraint the constraint/rule that is broken, e.g.: ConstraintErrors::LENGTH_MIN()
-     * @param JsonPointer|null $path
-     * @param array            $more       more array elements to add to the error
+     * @param ConstraintError $constraint the constraint/rule that is broken, e.g.: ConstraintErrors::LENGTH_MIN()
+     * @param array           $more       more array elements to add to the error
      */
-    public function addError(ConstraintError $constraint, ?JsonPointer $path = null, array $more = []);
+    public function addError(ConstraintError $constraint, ?JsonPointer $path = null, array $more = []): void;
 
     /**
      * checks if the validator has not raised errors
-     *
-     * @return bool
      */
-    public function isValid();
+    public function isValid(): bool;
 
     /**
      * invokes the validation of an element
      *
      * @abstract
      *
-     * @param mixed            $value
-     * @param mixed            $schema
-     * @param JsonPointer|null $path
-     * @param mixed            $i
+     * @param mixed $value
+     * @param mixed $schema
+     * @param mixed $i
      *
      * @throws \JsonSchema\Exception\ExceptionInterface
      */
-    public function check(&$value, $schema = null, ?JsonPointer $path = null, $i = null);
+    public function check(&$value, $schema = null, ?JsonPointer $path = null, $i = null): void;
 }