+++ src/JsonSchema/Constraints/SchemaConstraint.php
@@ -31,7 +31,7 @@ class SchemaConstraint extends Constraint
     /**
      * {@inheritdoc}
      */
-    public function check(&$element, $schema = null, ?JsonPointer $path = null, $i = null)
+    public function check(&$element, $schema = null, ?JsonPointer $path = null, $i = null): void
     {
         if ($schema !== null) {
             // passed schema