+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Used PHPStan's int-mask-of<T> type where applicable ([#779](https://github.com/jsonrainbow/json-schema/pull/779))
+- Fixed some PHPStan errors ([#781](https://github.com/jsonrainbow/json-schema/pull/781))
 
 ## [6.1.0] - 2025-02-04
 ### Added